Change 307 response description to fix OpenAPI error
During creation of the java script client from the OpenAPI file an error
was thrown because the unsupported field "message" was set for the 307
response:
```
Exception in thread "main" org.openapitools.codegen.SpecValidationException: There were issues with the specification. The option can be disabled via validateSpec (Maven/Gradle) or --skip-validate-spec (CLI).
 | Error count: 4, Warning count: 0
Errors:
	-attribute paths.'/projects/{project_id}/jobs/{job_id}/actions/{action_id}'(get).responses.307.message is unexpected
	-attribute paths.'/projects/{project_id}/jobs'(post).responses.307.message is unexpected
	-attribute paths.'/projects/{project_id}/services/{service_id}/actions/{action_id}'(get).responses.307.message is unexpected
	-attribute paths.'/projects/{project_id}/files/{file_key}/actions/{action_id}'(get).responses.307.message is unexpected
```
This is fixed by using the field "description" instead of "message"
which is part of the OpenAPI spec.