Extract project id from service metadata in map_service_to_extension (#13)
* Extract project id from service metadata in map_service_to_extension

The project id does not need to be passed to the `map_service_to_extension`
function because it can be retrieved from the service metadata. This fixes a
bug in `list_extensions` where the wrong project_id is passed for global
extensions.

* Improve integration test for list_extensions

Modify the list_extensions test so it reproduces the bug fixed by the
previous commit.