Connect backend to all docker networks of the installation on start up. (#19)
For each service that is started, a dedicated docker network is created. It includes only the backend and service containers. That way the service is isolated on a network level and can only communicate with the backend. After the backend container is recreated (e.g. updating the image), it is no longer part of the networks. This commit checks all networks at server startup and adds the backend back to the networks that belong to the installation.