Add API Token to each service/job which allows access to itself (#27)
Some service require a token that allows access to the service itself. For example, a ML workspace service needs this token to build the SSH setup command (the token has to be included in a curl call that downloads the setup script from the workspace).
This PR adds the environment variable CONTAXY_API_TOKEN to each service which includes a token that provides read access to this service. For this, the AuthManager needs to be passed down through several function calls. In the get_default_environment_variables function it is then used to generate the token.